src_test_process/requirements.txt
boto3==1.34.7

fastapi==0.105.0

loguru==0.7.2


psycopg2-binary==2.9.9


python-dotenv==1.0.0
sqlalchemy==2.0.23

uvicorn==0.25.0
python-multipart

pg8000==1.29.8

cloud-sql-python-connector==1.2.4

Werkzeug==2.3.7